refactor: rename `vercelStegaCleanAll` to `stegaClean`
