+++ src/stega/stegaClean.ts
